+++ app/src/main/kotlin/com/fernandocejas/sample/features/thridpartyinformation/ThirdPartyFragment.kt
@@ -204,9 +204,8 @@ class ThirdPartyFragment : BaseFragment(), View.OnClickListener {
 
                 var datos = idThird.split(";")
                 datos.get(0)
-//                oswaldo;lopez;jfjakd;659498;hzduir;Si;;;hyundai;accent;ifisid;2005;plomo;
                 var vehiculo = Vehiculo(datos.get(8),datos.get(9),datos.get(10),datos.get(11).toInt(),datos.get(12))
-                var tercero = Terceros(datos.get(0),datos.get(1),datos.get(2),datos.get(3).toInt(),datos.get(4),datos.get(5),datos.get(6),datos.get(7),"","","","","","","","","",vehiculo)
+                var tercero = Terceros(datos.get(0),datos.get(1),datos.get(2),datos.get(3),datos.get(4),datos.get(5),datos.get(6),datos.get(7),"","","","","","","","","",vehiculo)
 
                 setDataThird(tercero)
             }
@@ -296,7 +295,7 @@ class ThirdPartyFragment : BaseFragment(), View.OnClickListener {
                 color.text.toString())
 
         var terceros = Terceros(firstName.text.toString(),surname.text.toString(),
-                rut.text.toString(),telefono.text.toString().toInt(),email.text.toString(),
+                rut.text.toString(),telefono.text.toString(),email.text.toString(),
                 spinner.selectedItem.toString(),idPath,driverLicencePath,"","","","","","","","","",vehiculo)
 
         //si idThird esta vacio se actualiza los datos del tercero sino agregar un tercero nu evo
